Improves side menu behavior and usability
Enhances the side menu's user experience by automatically closing it upon selecting a Pokémon.

This change replaces the toggle function with a dedicated close function, ensuring the menu closes when a user selects a Pokémon, thus providing a smoother navigation flow. Also, it improves CSS classes related to the menu to improve the menu's behaviour. Finally, it hides the scroll bar.